+++ plugins/src/main/kotlin/io/piano/android/configuration/CommonAndroidConfigurationPlugin.kt
@@ -32,8 +32,8 @@ class CommonAndroidConfigurationPlugin : Plugin<Project> {
             buildToolsVersion = AndroidConfig.androidBuildTools
             compileSdkVersion(AndroidConfig.androidCompileSdk)
             defaultConfig {
-                minSdkVersion(AndroidConfig.androidMinSdk)
-                targetSdkVersion(AndroidConfig.androidTargetSdk)
+                minSdk = AndroidConfig.androidMinSdk
+                targetSdk = AndroidConfig.androidTargetSdk
                 testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
             }
             compileOptions {